Move components into subdirectories
