+++ src/components/CourseList/CourseItem.scss
