+++ src/pages/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import NavBar from '../components/NavBar';
+import NavBar from '../components/Navigation/NavBar';
 
 
 const App = React.createClass({
